buffer: add advance(unsigned) back
advance(unsigned) and seek(int) were replaced with advance(size_t> and
seek(ssize_t) in 053bfa6. this change broke the backward compatibility
of librados/librbd's ABI.

- deprecates advance(unsigned) and seek(int)
- update the callers to use the prefered versions.

Fixes: http://tracker.ceph.com/issues/17809

+++ src/compressor/snappy/SnappyCompressor.h
@@ -45,7 +45,7 @@ class CEPH_BUFFER_API BufferlistSource : public snappy::Source {
   }
   void Skip(size_t n) override {
     assert(n <= remaining);
-    pb.advance(n);
+    pb.advance(static_cast<ssize_t>(n));
     remaining -= n;
   }
